feat: download url operator (#16)
Because

In VDP pipeline builder, we only support use base64 to transfer the
unstructured data (e.g. image, audio). In some scenarios, the user may
only want to use a url link to the unstructured data as input. So we
need to find a way to download the data from url into the pipeline.

This commit

- Adds a download url operator,
- Unit tested
- Fix and refactor integration testing
- Closes https://github.com/instill-ai/community/issues/399