feat: auto-generate operator README.mdx (#65)
# What is the purpose of this PR?

![CleanShot 2024-02-23 at 12 31
08](https://github.com/instill-ai/operator/assets/3977183/efc49ebc-4023-4845-9ed8-d79bb44eb82a)

Because

- Component docs get outdated easily and keeping them up-to-date is a
(mostly) mechanical task.

This PR

- Leverages `compogen readme` (see
https://github.com/instill-ai/component/pull/59) to generate
automatically a `README.mdx` file for each operator.

## Excluded operators

- `start` and `end`, as their structure is special and shouldn't be
treated as operators
- `image`, as `$ref` schema isn't supported by `compogen` yet

## Next steps / improvements

The main goal of `compogen` is minimizing the maintenance cost of the
component docs at
[instill.tech](https://www.instill.tech/docs/v0.10.0-beta/vdp/operators/json).
That's why the output is an MDX document. However, having a Markdown
version of the document would be beneficial for developers using this
repository, removing the need to access an external website (or offering
the rendered Markdown in GitHub at the root of the component).
Transforming MDX to Markdown should be
[feasible](https://www.npmjs.com/package/mdx-to-md) but I didn't manage
to automate the process.

It would be interesting to add input / output code samples, either by
adding an example field at the task level in `tasks.json` or by checking
if all the required fields have an `example` field.

If this proves useful and production-ready, we can implement an GitHub
action that generates the documents automatically when changes are
introduced in `tasks.json` or `definitions.json`. We can even push these
changes to the website repo.